requirements.txt
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
selenium==4.38.0
black==25.9.0
flake8==7.3.0
isort==7.0.0
allure-pytest==2.15.2
selene==2.0.0rc9
python-dotenv==1.2.1